+++ src/backend/src/infrastructure/database/queries/games/listGames.js
