+++ src/backend/src/infrastructure/database/queries/users/getUser.js
